some tweaks
- reduced file size of 'favicon.ico' to decrease load time
- added some testing resources for future compliance with web standards
- researched testing CI (continuous integration) to immediately report errors
- adjusted route calls for cleaner code
- more comments to reduce my loss of mind
