+++ README.md
@@ -3,6 +3,7 @@
 a place to build cool stuff and see awesome things
 
 # Stylings
+
 <table border=3>
 <header>
 <th>
@@ -26,6 +27,9 @@ a place to build cool stuff and see awesome things
 </tr>
 </table>
 
+## Testing
+[html validator](https://validator.w3.org/nu/?doc=https%3A%2F%2Flimitedprintworld.com%2F)
+[css validator](https://jigsaw.w3.org/css-validator/validator?uri=https%3A%2F%2Flimitedprintworld.com%2F&profile=css3svg&usermedium=all&warning=1&vextwarning=&lang=en)
 
 
 <!--
