+++ routes/html.py
@@ -1,12 +1,15 @@
-from fastapi import APIRouter, Request, Form
-from fastapi.responses import HTMLResponse, FileResponse
-from fastapi.templating import Jinja2Templates
 from pathlib import Path
-
+from enum import Enum
 import os,sys,time,json
 
 
-from enum import Enum
+from routes.forms import router as forms_router
+
+from fastapi import APIRouter, Request, Form
+from fastapi.responses import HTMLResponse, FileResponse
+from fastapi.templating import Jinja2Templates
+
+
 
 class FormName(str, Enum):
     generalQuery = "generalQuery"
@@ -47,9 +50,9 @@ async def index(request: Request):
         return decorator*level + title.title() + decorator*level
     creditsList = [
         header("Graphical User Interface (GUI)"),
-        "Sound Effect by <a href="https://pixabay.com/users/fronbondi_skegs-23154649/?utm_source=link-attribution&utm_medium=referral&utm_campaign=music&utm_content=359401">Gavin Mogensen</a> from <a href="https://pixabay.com//?utm_source=link-attribution&utm_medium=referral&utm_campaign=music&utm_content=359401">Pixabay</a>",
+        'Sound Effect by <a href="https://pixabay.com/users/fronbondi_skegs-23154649/?utm_source=link-attribution&utm_medium=referral&utm_campaign=music&utm_content=359401">Gavin Mogensen</a> from <a href="https://pixabay.com//?utm_source=link-attribution&utm_medium=referral&utm_campaign=music&utm_content=359401">Pixabay</a>',
         header("sounds, effect, and ambience",),
-        "",
+        '',
     ]
 
     return templates.TemplateResponse("credits.html", {"request": request,\
@@ -70,23 +73,3 @@ async def submit_email(request: Request, email: str = Form(...)):
         "message": email
     })
 
-@router.get("/forms", response_class=HTMLResponse)
-async def get_forms(request: Request, ):
-
-    links = []
-    
-    for page in os.listdir(root/"templates/forms"): 
-        pageName=page.split(".")[0]
-        print("#"*6)
-
-        print(pageName, page)
-        print("#"*6)
-
-        links.append(pageName)
-    return templates.TemplateResponse("forms.html", {"request": request,"form_collection": links.copy() })
-@router.get("/forms/{form_name}", response_class=HTMLResponse)
-async def get_forms(request: Request, form_name:FormName):
-    forms = FormName
-    forms.generalQuery
-    if form_name is forms.generalQuery:
-        return forms.TemplateResponse(f"{form_name.generalQuery}.html", {"request": request,})
