+++ templates/credits.html
@@ -10,14 +10,17 @@
     
     
     <link rel="shortcut icon" href="../static/favicon.ico" type="image/x-icon">
-   
+    
 </head>
 <body class="film-grain">
 
 <NAV>
     <ul>
         <li>
-            <a href="/"><img src="/static/favicon.ico" class="icon" alt="<- Homepage"></a>
+            <div class="item-card">
+
+                <a href="/"><img src="/static/favicon.ico" class="icon" alt="<- Homepage"></a>
+            </div>
         </li>
         <li>Homepage</li>
     </ul>
