+++ templates/forms/generalQuery.html
@@ -8,6 +8,7 @@
     <link rel="stylesheet" href="/static/css/generalQuery.css">
 </head>
 <body>
+
 <h1>General Inquirey</h1>
 <h2>Limited Print World </h2>
 <h3>3D Printing Services</h3>
@@ -17,7 +18,7 @@ we're a small team of dedicated individuals seeking to provide some sustainable 
     </p>
 <p>
 </p>
-    <p>
+<p>
 It takes time to complete our multifacitude of duties and respectfully respond to letters.
 </p>
 <p>
@@ -25,7 +26,6 @@ We look forward to hearing from everyone good or bad for any feedback
 
 </p>
 <div class="item-card">
-
     <form action="/submit-email" method="post" enctype="multipart/form-data" autocomplete="off" >
         <label for="name">Your Name (required):</label>
         <input type="text" id="name" name="name" required placeholder="Jane Doe"><br>
