+++ templates/index.html
@@ -20,7 +20,7 @@
         <h1 >Limited Print World</h1>
         <h1 style="color: maroon;">( ~_~ Under Construction ~_~ )</h1>
         <p style="font-style: italic; font-size: 200%;"><strong>(~Version_{{ v_id }})</strong></p>
-        <img src="/static/img/favicon.ico" alt="favicon" class="icon">
+        <img src="/static/img/favicon.ico" alt="favicon" class="icon" fetchpriority=high>
         <p style="font-size: 225%;" >Miniatures & Decorations for the Discerning Collector</p>
     
     <!-- Ambient background sound -->
